avoid logging error when error callback supplied
This should fix the following issue:

While trying to sync the app with an error callback, we should assume that the callback should handle and log the error.

In my case, I receive an error that is logged to Sentry whenever an update is not available. Passing an error callback doesn't solve it. Even wrapping it inside try-catch.